refactor: delete some unnesessary files
