+++ src/js/markup/create-cart-items-list.ts
@@ -1,6 +1,11 @@
 import { DataType } from "../types/types";
+import { EMPTY_CART } from "../utils/const";
 
-export function createCartItem({id, title, description, price, stock, thumbnail, amount}: DataType) {
+export function createCartItemsList(products: DataType[]) {
+    return `<ul class="cart__products-list cart-products-list">${products.length ? products.map((product: DataType) => createCartItem(product)).join('') : EMPTY_CART}</ul>`
+}
+
+export function createCartItem({ id, title, description, price, stock, thumbnail, amount }: DataType) {
     return `<li class="cart-products-list__item cart-item">
               <p class="cart-item__title">${title}</p>
               <img
