--- src/js/markup/create-footer.ts
@@ -1,5 +1,0 @@
-export function createFooter() {
-   return `<footer class="footer">
-   <span>Online-Store, 2023</span>
-   </footer>`
-}
