+++ src/js/pages/create-product.ts
@@ -1,7 +1,6 @@
 import { createBreadcrumbs } from "../markup/create-breadcrumbs";
-import { createHeader } from "../markup/create-header";
-import { createProductExample } from "../markup/create-product-example";
-import { createFooter } from "../markup/create-footer";
+import { createHeader, createFooter } from "../markup/create-header-footer";
+import { createProductExample } from "../markup/create-product-markup";
 import { createSearchUrl, getSearchParams, hashListener } from "../utils/utils";
 import { getNodes } from "../markup/get-nodes";
 import { LS_KEYS, routes, SEARCH_KEYS } from "../utils/const";
@@ -16,9 +15,9 @@ export default function CreateProduct() {
   const productId = Number(location.slice(-3).match(/[0-9]/gi)?.join(''));
   const filtredData = mocks.filter((product) => product.id === productId);
   const isInCart = getFromLocalStorage(LS_KEYS.cart);
-  const ids = isInCart.map(({id}:{id: number}) => id);
+  const ids = isInCart.map(({ id }: { id: number }) => id);
   const { searchParams } = getSearchParams();
-  
+
   const body = document.querySelector(".page") as HTMLBodyElement;
   body.innerHTML = `${createHeader()}<main>${createBreadcrumbs(filtredData)}${createProductExample(filtredData)}</main>${createFooter()}`;
